backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
mysqlclient==2.2.4
openpyxl==3.1.5
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
